+++ Render/ISwapChain.h
@@ -11,7 +11,7 @@ namespace Lightning
 			virtual INTERFACECALL ~ISwapChain() = default;
 			//preset the current back buffer to display
 			virtual bool INTERFACECALL Present() = 0;
-			virtual void INTERFACECALL Resize(std::uint32_t width, std::uint32_t height) = 0;
+			virtual void INTERFACECALL Resize(std::size_t width, std::size_t height) = 0;
 			virtual std::size_t INTERFACECALL GetMultiSampleCount()const = 0;
 			virtual std::size_t INTERFACECALL GetMultiSampleQuality()const = 0;
 			virtual RenderFormat INTERFACECALL GetRenderFormat()const = 0;
